test(e2e): improve directory structure of some cases
