--- e2e/cases/performance/basic/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@rsbuild/tsconfig/base",
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "baseUrl": "./",
-    "outDir": "./dist"
-  },
-  "include": ["src"]
-}
