+++ e2e/cases/performance/dns-prefetch/rsbuild.config.ts
@@ -1,0 +1,8 @@
+import { pluginReact } from '@rsbuild/plugin-react';
+
+export default {
+  plugins: [pluginReact()],
+  performance: {
+    dnsPrefetch: ['http://aaaa.com'],
+  },
+};
