+++ e2e/cases/performance/dns-prefetch/src/App.jsx
