+++ e2e/cases/performance/dns-prefetch/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App';
+import App from './App.jsx';
 
 const container = document.getElementById('root');
 if (container) {
