+++ e2e/cases/performance/preconnect/rsbuild.config.ts
@@ -1,0 +1,16 @@
+import { pluginReact } from '@rsbuild/plugin-react';
+
+export default {
+  plugins: [pluginReact()],
+  performance: {
+    preconnect: [
+      {
+        href: 'http://aaaa.com',
+      },
+      {
+        href: 'http://bbbb.com',
+        crossorigin: true,
+      },
+    ],
+  },
+};
