+++ e2e/cases/performance/preconnect/src/App.jsx
@@ -1,0 +1,3 @@
+const App = () => <div id="test">Hello Rsbuild!</div>;
+
+export default App;
