+++ e2e/cases/performance/resource-hints/index.test.ts
