+++ e2e/cases/performance/resource-hints/src/env.d.ts
