+++ e2e/cases/performance/resource-hints/src/page1/App.tsx
