+++ e2e/cases/performance/resource-hints/src/page1/index.ts
