+++ e2e/cases/performance/resource-hints/src/page2/App.tsx
