+++ e2e/cases/performance/resource-hints/src/test.ts
