+++ e2e/cases/performance/resource-hints/tsconfig.json
