+++ e2e/cases/performance/split-chunk-all-in-one/rsbuild.config.ts
@@ -1,0 +1,10 @@
+import { pluginReact } from '@rsbuild/plugin-react';
+
+export default {
+  plugins: [pluginReact()],
+  performance: {
+    chunkSplit: {
+      strategy: 'all-in-one',
+    },
+  },
+};
