+++ e2e/cases/performance/split-chunk-by-module/src/App.jsx
