+++ e2e/cases/performance/split-chunk-by-module/src/index.js
