+++ e2e/cases/performance/split-chunk-single-vendor/src/App.jsx
